chore(kalert): resize observer infinite loop [KHCP-12603]
